SetPivotFromPixel has been implemented for OpenGL ES, as glReadPixels does not follow the same specifications.
- draw the scene in a grayscale representing the depth value
accessible only from the fragment shader. To do so, uses a vector
containing all of the programs used to render the scene.
The fragment shaders of all these programs is replaced by one
drawing in a grayscale, then the scene is rendered in an FBO that
will not be displayed. Then glReadPixels reads the color data, which
is then converted in a depth value, and the original fragment shaders
are set back up.